Remove ConcurrencyOptionProvider and merge into SyncTriggerOptionProvider
