--- src/WebJobs.Script.WebHost/IExtensionsOptionProvider.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-using System.Collections.Generic;
-using Newtonsoft.Json.Linq;
-
-namespace Microsoft.Azure.WebJobs.Script.WebHost
-{
-    public interface IExtensionsOptionProvider
-    {
-        Dictionary<string, JObject> GetExtensionOptions();
-    }
-}
