+++ src/WebJobs.Script.WebHost/Management/IFunctionsSyncManager.cs
@@ -8,10 +8,6 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
 {
     public interface IFunctionsSyncManager
     {
-        IExtensionsOptionProvider ExtensionsOptionProvider { set; }
-
-        IConcurrencyOptionProvider ConcurrencyOptionProvider { set; }
-
         /// <summary>
         /// Sync function triggers with Antares infrastructure.
         /// </summary>
